DO NOT INCORPORATE TO MAIN BRANCH!!!
This is an upload for you to test the targeting system, as well as to see what has been changed, and how it will affect what you guys are working on. I'm still doing some fun things with custom localization, and a few events are still broken / unfinished.

If you have any critiques, concerns, or suggestions, I'd also be happy to hear them.