Finally Done (Mostly)
I'm happy with this push as a beta build. Please give it a try and tell me what you think, any concerns, criticisms, or ideas for improvement. I took some liberties with the localization, so let me know if you don't like the style or some of the concepts I introduced.