Custom Coming of Age Event (Read Desc)
Important Note: I was unable to untangle the vanilla method of handing out traits with the time I have, so I created a custom event which hands out the education traits based on three factors - magic potential, intellect traits, and whether the kid has the curious trait. This should work as a quick-fix, but it could use some improvement. Also, I added an interaction as a self-help fix for players that are stuck with grown-ass wards. They need debug mode on and they just need to right click the ward, and they will be removed (Yes. Grown-ass wards were a thing because the event wasn't removing them prior to this update)