Modifier Update With Custom Icons + Effect Loc
This is a bag of goodies push. I changed a lot of things to make the consolidated modifiers work. In addition, I gave each aura its own modifier! Let me know if you don't like any of them or think that the icon should be something else. I also went through and made a few tweaks to localization for effects, so that the player doesn't have to read the description to know why they can't cast the spell. Lastly, I removed traits from the ruler designer.