Corrected Issue With Legacy Modifiers Not Being Removed
Just moved the effect which removes the legacy modifiers from the remove_aura_effect to the first page of the spellbook events. This removes all instances of the old modifiers from characters if they exists, and should be removed after the next update. It's just a save game compatability measure.