Spell target select update
Change maximum displayed spell targets to 255 so they don't overlap each other at the top if there's too many.
Not sure if `maxverticalslots` is a byte or a larger data type, but I would figure a byte is all they *should* need, so maybe they limited it. If 255 is not enough, we'll adapt